fix fork control for unit test
