+++ chainbase/src/main/java/org/tron/core/capsule/ReceiptCapsule.java
@@ -5,7 +5,7 @@ import lombok.Getter;
 import lombok.Setter;
 import org.tron.common.utils.Commons;
 import org.tron.common.utils.DBConfig;
-import org.tron.common.utils.ForkUtils;
+import org.tron.common.utils.ForkController;
 import org.tron.common.utils.Sha256Hash;
 import org.tron.common.utils.StringUtil;
 import org.tron.core.Constant;
@@ -113,7 +113,7 @@ public class ReceiptCapsule {
    * payEnergyBill pay receipt energy bill by energy processor.
    */
   public void payEnergyBill(DynamicPropertiesStore dynamicPropertiesStore,
-      AccountStore accountStore, ForkUtils forkUtils, AccountCapsule origin, AccountCapsule caller,
+      AccountStore accountStore, ForkController forkController, AccountCapsule origin, AccountCapsule caller,
       long percent, long originEnergyLimit, EnergyProcessor energyProcessor, long now)
       throws BalanceInsufficientException {
     if (receipt.getEnergyUsageTotal() <= 0) {
@@ -121,13 +121,13 @@ public class ReceiptCapsule {
     }
 
     if (Objects.isNull(origin) && dynamicPropertiesStore.getAllowTvmConstantinople() == 1) {
-      payEnergyBill(dynamicPropertiesStore, accountStore, forkUtils, caller,
+      payEnergyBill(dynamicPropertiesStore, accountStore, forkController, caller,
           receipt.getEnergyUsageTotal(), energyProcessor, now);
       return;
     }
 
     if (caller.getAddress().equals(origin.getAddress())) {
-      payEnergyBill(dynamicPropertiesStore, accountStore, forkUtils, caller,
+      payEnergyBill(dynamicPropertiesStore, accountStore, forkController, caller,
           receipt.getEnergyUsageTotal(), energyProcessor, now);
     } else {
       long originUsage = Math.multiplyExact(receipt.getEnergyUsageTotal(), percent) / 100;
@@ -138,7 +138,7 @@ public class ReceiptCapsule {
       long callerUsage = receipt.getEnergyUsageTotal() - originUsage;
       energyProcessor.useEnergy(origin, originUsage, now);
       this.setOriginEnergyUsage(originUsage);
-      payEnergyBill(dynamicPropertiesStore, accountStore, forkUtils,
+      payEnergyBill(dynamicPropertiesStore, accountStore, forkController,
           caller, callerUsage, energyProcessor, now);
     }
   }
@@ -155,7 +155,7 @@ public class ReceiptCapsule {
   }
 
   private void payEnergyBill(
-      DynamicPropertiesStore dynamicPropertiesStore, AccountStore accountStore, ForkUtils forkUtils,
+      DynamicPropertiesStore dynamicPropertiesStore, AccountStore accountStore, ForkController forkController,
       AccountCapsule account,
       long usage,
       EnergyProcessor energyProcessor,
@@ -167,7 +167,7 @@ public class ReceiptCapsule {
     } else {
       energyProcessor.useEnergy(account, accountEnergyLeft, now);
 
-      if (forkUtils.pass(ForkBlockVersionEnum.VERSION_3_6_5) &&
+      if (forkController.pass(ForkBlockVersionEnum.VERSION_3_6_5) &&
           dynamicPropertiesStore.getAllowAdaptiveEnergy() == 1) {
         long blockEnergyUsage =
             dynamicPropertiesStore.getBlockEnergyUsage() + (usage - accountEnergyLeft);
