+++ chainbase/src/main/java/org/tron/core/db/TransactionTrace.java
@@ -15,7 +15,7 @@ import org.tron.common.runtime.Runtime;
 import org.tron.common.runtime.vm.DataWord;
 import org.tron.common.utils.Commons;
 import org.tron.common.utils.DBConfig;
-import org.tron.common.utils.ForkUtils;
+import org.tron.common.utils.ForkController;
 import org.tron.common.utils.Sha256Hash;
 import org.tron.common.utils.WalletUtil;
 import org.tron.common.utils.DecodeUtil;
@@ -66,7 +66,7 @@ public class TransactionTrace {
 
   private Runtime runtime;
 
-  private ForkUtils forkUtils;
+  private ForkController forkController;
 
   @Getter
   private TransactionContext transactionContext;
@@ -98,8 +98,8 @@ public class TransactionTrace {
     this.receipt = new ReceiptCapsule(Sha256Hash.ZERO_HASH);
     this.energyProcessor = new EnergyProcessor(dynamicPropertiesStore, accountStore);
     this.runtime = runtime;
-    this.forkUtils = new ForkUtils();
-    forkUtils.init(dynamicPropertiesStore);
+    this.forkController = new ForkController();
+    forkController.init(storeFactory.getChainBaseManager());
   }
 
   public TransactionCapsule getTrx() {
@@ -227,7 +227,7 @@ public class TransactionTrace {
     AccountCapsule origin = accountStore.get(originAccount);
     AccountCapsule caller = accountStore.get(callerAccount);
     receipt.payEnergyBill(
-        dynamicPropertiesStore, accountStore, forkUtils,
+        dynamicPropertiesStore, accountStore, forkController,
         origin,
         caller,
         percent, originEnergyLimit,
